Truly fix #251
